+++ CHANGES.rst
@@ -4,6 +4,10 @@ Change history
 2.0.1 (unreleased)
 ------------------
 
+- Change pep3101 logic. No more false positives on log
+  strings.
+  [do3cc]
+
 - Clean tests output.
   Fixes https://github.com/plone/plone.recipe.codeanalysis/issues/122
   [gforcada]